+++ src/lib/gadgets/range-check.unit-test.ts
@@ -1,5 +1,5 @@
 import { mod } from '../../bindings/crypto/finite_field.js';
-import { Field } from '../field.js';
+import { Field } from '../../lib/core.js';
 import { ZkProgram } from '../proof_system.js';
 import {
   Spec,
@@ -7,10 +7,20 @@ import {
   equivalentAsync,
   field,
 } from '../testing/equivalent.js';
-import { Random } from '../testing/random.js';
+import { Random } from '../testing/property.js';
 import { Gadgets } from './gadgets.js';
 
+let maybeUint64: Spec<bigint, Field> = {
+  ...field,
+  rng: Random.map(Random.oneOf(Random.uint64, Random.uint64.invalid), (x) =>
+    mod(x, Field.ORDER)
+  ),
+};
+
 // TODO: make a ZkFunction or something that doesn't go through Pickles
+// --------------------------
+// RangeCheck64 Gate
+// --------------------------
 
 let RangeCheck64 = ZkProgram({
   name: 'range-check-64',
@@ -26,16 +36,7 @@ let RangeCheck64 = ZkProgram({
 
 await RangeCheck64.compile();
 
-let maybeUint64: Spec<bigint, Field> = {
-  ...field,
-  rng: Random.map(Random.oneOf(Random.uint64, Random.uint64.invalid), (x) =>
-    mod(x, Field.ORDER)
-  ),
-};
-
-// do a couple of proofs
 // TODO: we use this as a test because there's no way to check custom gates quickly :(
-
 await equivalentAsync({ from: [maybeUint64], to: boolean }, { runs: 3 })(
   (x) => {
     if (x >= 1n << 64n) throw Error('expected 64 bits');